+++ Models/FrameSortField.cs
@@ -2,6 +2,7 @@ namespace blink_o_mat.Models;
 
 public enum FrameSortField
 {
+    Score,
     ObservationDate,
     Fwhm,
     FwhmArcsec,